5_var Bruzgin (#12)
 исправление первой лабы вар 5
